+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 2018/02/18
+
+* change `fund-in.sh` to `pay_fundin.sh`
+
+```bash
+pay_fundin.sh <pay to fund-in tx(satoshi)> <pay to our channel(satoshi)> <pay to their channel(satoshi)>
+```
+
 ## 2018/02/17
 
 * deprecate `node.conf`